DF-10193 fix after merge development
Former-commit-id: 8637a3b51b0124db9c28195fa58ab3d0e7ecce3f

+++ Sources/Internal/FileSystem/DynamicMemoryFile.cpp
@@ -153,13 +153,13 @@ bool DynamicMemoryFile::IsEof() const
     return isEof;
 }
 
-bool DynamicMemoryFile::Truncate(int32 size)
+bool DynamicMemoryFile::Truncate(uint64 size)
 {
     if (!(fileAttributes & File::WRITE))
         return false;
 
     data.resize(size);
-    currentPtr = Min(currentPtr, uint32(size));
+    currentPtr = Min(currentPtr, size);
     isEof = (currentPtr == size);
 
     return true;
